pyproject.toml
[project]
name = "ai-scenario-modeling"
version = "1.0.0"
description = "AI-Enhanced Data Analytics & Scenario Modeling — Portfolio Project"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "plotly>=5.18.0",
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "nbformat>=5.9.0",
    "ipywidgets>=8.1.0",
    "kaleido>=0.2.1",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["shared*"]

[tool.ruff]
line-length = 100
target-version = "py310"
